campaign/requirements.txt
flask>=2.3.0
pyyaml>=6.0
jinja2>=3.1.0
requests>=2.28.0
